Fixed DB calls for MySQL

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -26,7 +26,7 @@ class CreateUsersTable extends Migration
             $table->rememberToken();
             $table->timestamps();
             $table->string('theme')->nullable();
-            $table->unsignedBigInteger('auth_as')->nullable();
+            $table->unsignedInteger('auth_as')->nullable();
         });
     }
